refactor!: move `close_on_esc` to `bevy_dev_tools`
I chose to remove references to `close_on_esc` in examples, since `bevy_dev_tools` is not enabled by default.
