+++ crates/bevy_dev_tools/src/close_on_esc.rs
@@ -1,0 +1,32 @@
+use bevy_ecs::prelude::*;
+use bevy_input::{keyboard::KeyCode, ButtonInput};
+use bevy_window::Window;
+
+/// Close the focused window whenever the escape key (<kbd>Esc</kbd>) is pressed
+///
+/// This is useful for examples or prototyping.
+///
+/// # Example
+///
+/// ```no_run
+/// # use bevy_app::prelude::*;
+/// # use bevy_dev_tools::close_on_esc;
+/// #
+/// App::new()
+///     .add_systems(Update, close_on_esc);
+/// ```
+pub fn close_on_esc(
+    mut commands: Commands,
+    focused_windows: Query<(Entity, &Window)>,
+    input: Res<ButtonInput<KeyCode>>,
+) {
+    for (window, focus) in focused_windows.iter() {
+        if !focus.focused {
+            continue;
+        }
+
+        if input.just_pressed(KeyCode::Escape) {
+            commands.entity(window).despawn();
+        }
+    }
+}
