+++ examples/games/breakout.rs
@@ -75,7 +75,7 @@ fn main() {
                 // `chain`ing systems together runs them in order
                 .chain(),
         )
-        .add_systems(Update, (update_scoreboard, bevy::window::close_on_esc))
+        .add_systems(Update, update_scoreboard)
         .run();
 }
