+++ examples/window/multiple_windows.rs
@@ -7,7 +7,6 @@ fn main() {
         // By default, a primary window gets spawned by `WindowPlugin`, contained in `DefaultPlugins`
         .add_plugins(DefaultPlugins)
         .add_systems(Startup, setup_scene)
-        .add_systems(Update, bevy::window::close_on_esc)
         .run();
 }
